+++ src-docs/src/views/tabs/tabbed_content.js
@@ -13,7 +13,7 @@ class EuiTabsExample extends Component {
 
     this.tabs = [
       {
-        id: 'cobalt',
+        id: 'cobalt--id',
         name: 'Cobalt',
         content: (
           <Fragment>
@@ -32,7 +32,7 @@ class EuiTabsExample extends Component {
         ),
       },
       {
-        id: 'dextrose',
+        id: 'dextrose--id',
         name: 'Dextrose',
         content: (
           <Fragment>
@@ -49,7 +49,7 @@ class EuiTabsExample extends Component {
         ),
       },
       {
-        id: 'hydrogen',
+        id: 'hydrogen--id',
         name: 'Hydrogen',
         content: (
           <Fragment>
@@ -66,7 +66,7 @@ class EuiTabsExample extends Component {
         ),
       },
       {
-        id: 'monosodium_glutammate',
+        id: 'monosodium_glutammate--id',
         name: 'Monosodium Glutamate',
         content: (
           <Fragment>